refactor : user가 로그인했는지 확인하는 함수 userService로 이동
